Styling question:
I would like to be able to put a white box around items that don’t stand out.  I did it with specific code on two pages but am wondering if it could be an option in the section .css. (Bonus if we could assign it to individual columns, but I was able to do that by changing a symple code style in the rentals example below. ) Please take a look: